// src/app/models/magazines.ts
export class Magazine{
    id: number = 0;
    autor: string;
    titulo: string;
    edicion: string;
    descripcion: string;
    frecuencia_Actual: string;
    ejemplares: number;
    temas: string;
    palabrasClave: string;
    copias: number;
    disponibles: number;
    numPrestamo: number = 0;
}